Simplify iob plugin with a settings-aware calc() function
