+++ lib/plugins/index.js
@@ -29,7 +29,6 @@ function init() {
     , require('./cannulaage')()
     , require('./basalprofile')()
     , require('./upbat')()
-    , require('./pumpiob')()
   ];
 
   var serverDefaultPlugins = [
