+++ lib/plugins/iob.js
@@ -2,7 +2,6 @@
 
 var _ = require('lodash')
   , moment = require('moment')
-  , pumpiob = require('./pumpiob')
   , utils = require('../utils')();
 
 function init() {
@@ -12,60 +11,92 @@ function init() {
     , pluginType: 'pill-major'
   };
 
-  var bolusIobMixin = bolusiob(iob);
-  var pumpIobMixin = pumpiob(iob);
-
-  // TODO(@mddub|2015-11-02): these can use https://github.com/nightscout/dcalc
-  // when that package exposes these functions
-  iob.calcTotal = bolusIobMixin.calcTotal;
-  iob.calcTreatment = bolusIobMixin.calcTreatment;
-  iob.getPumpIOB = pumpIobMixin.getPumpIOB;
-
-  // Mix in the right iob mode at runtime, when we can get settings from the sandbox
-  function chooseMixin (sbx) {
-    return sbx.extendedSettings.source === 'pump' ? pumpIobMixin : bolusIobMixin;
+  function usePumpIOB (sbx) {
+    return sbx.extendedSettings.source === 'pump';
   }
+
+  iob.calc = function calc (sbx, time) {
+    // Choose calculation mode at runtime, when we can get extended settings from the sandbox
+    if (usePumpIOB(sbx)) {
+      return iob.getPumpIOB(sbx.data.pumpStatuses, sbx.extendedSettings, time);
+    } else {
+      return iob.calcTotal(sbx.data.treatments, sbx.data.profile, time);
+    }
+  };
+
   iob.setProperties = function setProperties (sbx) {
-    chooseMixin(sbx).setProperties(sbx);
-  };
-  iob.updateVisualisation = function updateVisualisation (sbx) {
-    chooseMixin(sbx).updateVisualisation(sbx);
-  };
-
-  return iob;
-}
-
-// Bolus IOB mixin
-function bolusiob (plugin) {
-  var biob = {};
-
-  biob.setProperties = function setProperties (sbx) {
-    sbx.offerProperty(plugin.name, function setIOB ( ) {
-      return biob.calcTotal(sbx.data.treatments, sbx.data.profile, sbx.time);
+    sbx.offerProperty('iob', function setIOB () {
+      return iob.calc(sbx, sbx.time);
     });
   };
 
-  biob.updateVisualisation = function updateVisualisation (sbx) {
-    var info = null;
+  iob.updateVisualisation = function updateVisualisation (sbx) {
+    var prop = sbx.properties.iob
+      , info
+      , value;
 
-    var prop = sbx.properties[plugin.name];
-
-    if (prop && prop.lastBolus) {
-      var when = moment(prop.lastBolus.mills).format('lll');
-      var amount = sbx.roundInsulinForDisplayFormat(Number(prop.lastBolus.insulin)) + 'U';
-      info = [{label: 'Last Bolus', value: amount + ' @ ' + when }];
+    if (usePumpIOB(sbx)) {
+      if (prop) {
+        value = prop.display + 'U';
+        if (prop.uploaderBattery) {
+          info = [{label: 'Uploader Battery', value: prop.uploaderBattery + '%'}];
+        }
+      } else {
+        value = '---';
+        info = [{label: 'Pump IOB', value: 'No recent data'}];
+      }
+    } else {
+      if (prop && prop.lastBolus) {
+        var when = moment(prop.lastBolus.mills).format('lll');
+        var amount = sbx.roundInsulinForDisplayFormat(Number(prop.lastBolus.insulin)) + 'U';
+        info = [{label: 'Last Bolus', value: amount + ' @ ' + when }];
+      }
+      value = sbx.roundInsulinForDisplayFormat(prop.display) + 'U';
     }
 
-    sbx.pluginBase.updatePillText(plugin, {
-      value: sbx.roundInsulinForDisplayFormat(prop.display) + 'U'
-      , label: 'IOB'
+    sbx.pluginBase.updatePillText(iob, {
+      label: 'IOB'
+      , value: value
       , info: info
     });
-
   };
 
-  // TODO(@mddub|2015-11-02): remove and use dcalc
-  biob.calcTotal = function calcTotal(treatments, profile, time) {
+  iob.getPumpIOB = function getPumpIOB (pumpStatuses, settings, time) {
+    settings = _.cloneDeep(settings);
+    settings.pumpRecency = settings.pumpRecency !== undefined ? parseInt(settings.pumpRecency, 10) : 10;
+    // 25 is the MiniMed Connect's threshold for LOW vs MEDIUM
+    settings.pumpBatteryPebble = settings.pumpBatteryPebble !== undefined ? parseInt(settings.pumpBatteryPebble, 10) : 25;
+
+    var lastPumpStatus = _.last(_.dropRightWhile(pumpStatuses, function (pumpStatus) {
+      return pumpStatus.mills > time;
+    }));
+
+    if (!lastPumpStatus || lastPumpStatus.activeInsulin === undefined) {
+      return null;
+    } else if (time - lastPumpStatus.mills > settings.pumpRecency * 60 * 1000) {
+      return null;
+    } else {
+      var iob = lastPumpStatus.activeInsulin;
+      var display = iob.toFixed(1);
+
+      var pebbleDisplay = display;
+      if (settings.pumpBatteryPebble && lastPumpStatus.conduitBatteryLevel <= settings.pumpBatteryPebble) {
+        pebbleDisplay += '*';
+      }
+
+      return {
+        iob: iob
+        , display: display
+        , displayLine: 'IOB: ' + display + 'U'
+        , pebbleDisplay: pebbleDisplay
+        , uploaderBattery: lastPumpStatus.conduitBatteryLevel
+      };
+    }
+  };
+
+  // TODO(@mddub|2015-11-02): remove these and use https://github.com/nightscout/dcalc
+  // when that package exposes these functions
+  iob.calcTotal = function calcTotal(treatments, profile, time) {
 
     var totalIOB = 0
       , totalActivity = 0;
@@ -80,7 +111,7 @@ function bolusiob (plugin) {
 
     _.each(treatments, function eachTreatment(treatment) {
       if (treatment.mills <= time) {
-        var tIOB = biob.calcTreatment(treatment, profile, time);
+        var tIOB = iob.calcTreatment(treatment, profile, time);
         if (tIOB.iobContrib > 0) {
           lastBolus = treatment;
         }
@@ -100,7 +131,7 @@ function bolusiob (plugin) {
     };
   };
 
-  biob.calcTreatment = function calcTreatment(treatment, profile, time) {
+  iob.calcTreatment = function calcTreatment(treatment, profile, time) {
 
     var dia = 3
       , sens = 0;
@@ -139,7 +170,7 @@ function bolusiob (plugin) {
 
   };
 
-  return biob;
+  return iob;
 }
 
 module.exports = init;
